clear 'npm install' artifacts from package.json
This project was likely started by running "npm install" to get its source.
In `package.json` there were fields, which are normally not part of package.json,
but npm adds them when it 'npm install's some dependency.
This commit removes those generated fields (which should not be present in normal project package.json).